requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
google-genai
httpx
rich
google-cloud-storage 
google-cloud-aiplatform